Fix protocol frame padding size
Since VLAN tags can be added or removed by any switch in the path of the
packet, ignore VLAN tags from padding size to ensure that the packet
will pass through all network devices.

Also change paddind method to Raw() due to issues with size of packets
in Dot1Q frames. Padding() is not counted in the frame size if the LLC()
layer if added before a Dot1Q tag.

Signed-off-by: Filipe Utzig <filipeutzig@gmail.com>